requirements/prod.txt
djangorestframework==3.16.1
python-dotenv==1.2.1